+++ js/get-random-counter.js
@@ -1,0 +1,19 @@
+
+getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+getRandomArrayElement = (array) =>array[getRandomInteger(0, array.length-1)];
+
+const createIdGenerator = () =>{
+  let lastGenerateId = 0;
+
+  return () => {
+    lastGenerateId += 1;
+    return lastGenerateId;
+  };
+};
